+++ app/build.gradle.kts
@@ -2,14 +2,15 @@ plugins {
     alias(libs.plugins.android.application)
     alias(libs.plugins.kotlin.android)
     alias(libs.plugins.kotlin.compose)
+    alias(libs.plugins.google.gms.google.services)
 }
 
 android {
-    namespace = "com.example.tictactoehunters"
+    namespace = "com.innovatelabz.tictactoehunters"
     compileSdk = 34
 
     defaultConfig {
-        applicationId = "com.example.tictactoehunters"
+        applicationId = "com.innovatelabz.tictactoehunters"
         minSdk = 24
         targetSdk = 34
         versionCode = 1
@@ -49,6 +50,7 @@ dependencies {
     implementation(libs.androidx.ui.graphics)
     implementation(libs.androidx.ui.tooling.preview)
     implementation(libs.androidx.material3)
+    implementation(libs.firebase.firestore)
     testImplementation(libs.junit)
     androidTestImplementation(libs.androidx.junit)
     androidTestImplementation(libs.androidx.espresso.core)
