+++ app/src/androidTest/java/com/innovatelabz/tictactoehunters/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.tictactoehunters
+package com.innovatelabz.tictactoehunters
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
