+++ app/src/main/java/com/innovatelabz/tictactoehunters/ui/theme/Theme.kt
@@ -1,6 +1,5 @@
-package com.example.tictactoehunters.ui.theme
+package com.innovatelabz.tictactoehunters.ui.theme
 
-import android.app.Activity
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme
