+++ app/src/main/java/com/innovatelabz/tictactoehunters/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.example.tictactoehunters.ui.theme
+package com.innovatelabz.tictactoehunters.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
